ai-service/requirements.txt
flask==3.0.3
flask-limiter==3.8.0
groq==0.9.0
redis==5.0.7
python-dotenv==1.0.1
pytest==8.2.2
requests==2.32.3
chromadb==0.5.3
sentence-transformers==3.0.1